Did code refactoring to the notifications
